+++ app/src/androidTest/java/com/example/shareroutine/FirebaseRealtimeDatabasePostTest.kt
@@ -3,8 +3,6 @@ package com.example.shareroutine
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.example.shareroutine.data.repository.PostRepositoryImpl
-import com.example.shareroutine.data.source.PostDataSource
-import com.example.shareroutine.data.source.realtime.PostDataSourceImplWithRealtime
 import com.example.shareroutine.domain.model.Post
 import com.example.shareroutine.domain.repository.PostRepository
 import com.example.shareroutine.domain.usecase.GetPostListUseCase
