+++ app/src/main/java/com/example/shareroutine/data/repository/PostRepositoryImpl.kt
@@ -2,7 +2,6 @@ package com.example.shareroutine.data.repository
 
 import com.example.shareroutine.data.mapper.PostMapper
 import com.example.shareroutine.data.source.PostDataSource
-import com.example.shareroutine.data.source.PostWithRoutineDataSource
 import com.example.shareroutine.data.source.realtime.State
 import com.example.shareroutine.di.IoDispatcher
 import com.example.shareroutine.domain.model.Post
@@ -15,7 +14,7 @@ import java.lang.Exception
 import javax.inject.Inject
 
 class PostRepositoryImpl @Inject constructor(
-    private val dataSource: PostWithRoutineDataSource,
+    private val dataSource: PostDataSource,
     @IoDispatcher private val ioDispatcher: CoroutineDispatcher
     ) : PostRepository {
     override suspend fun insert(post: Post) = withContext(ioDispatcher) {
