+++ app/src/main/java/com/example/shareroutine/data/source/PostDataSource.kt
@@ -1,13 +1,13 @@
 package com.example.shareroutine.data.source
 
 import com.example.shareroutine.data.source.realtime.State
-import com.example.shareroutine.data.source.realtime.model.RealtimeDBModelPost
+import com.example.shareroutine.data.source.realtime.model.RealtimeDBModelPostWithRoutine
 import kotlinx.coroutines.flow.Flow
 
 interface PostDataSource {
-    suspend fun insert(post: RealtimeDBModelPost)
-    suspend fun update(post: RealtimeDBModelPost)
-    suspend fun delete(post: RealtimeDBModelPost)
+    suspend fun insert(post: RealtimeDBModelPostWithRoutine)
+    suspend fun update(post: RealtimeDBModelPostWithRoutine)
+    suspend fun delete(post: RealtimeDBModelPostWithRoutine)
 
-    fun getAllPostList(): Flow<State<List<RealtimeDBModelPost>>>
+    fun getAllPostList(): Flow<State<List<RealtimeDBModelPostWithRoutine>>>
 }
