--- app/src/main/java/com/example/shareroutine/data/source/PostWithRoutineDataSource.kt
@@ -1,13 +1,0 @@
-package com.example.shareroutine.data.source
-
-import com.example.shareroutine.data.source.realtime.State
-import com.example.shareroutine.data.source.realtime.model.RealtimeDBModelPostWithRoutine
-import kotlinx.coroutines.flow.Flow
-
-interface PostWithRoutineDataSource {
-    suspend fun insert(post: RealtimeDBModelPostWithRoutine)
-    suspend fun update(post: RealtimeDBModelPostWithRoutine)
-    suspend fun delete(post: RealtimeDBModelPostWithRoutine)
-
-    fun getAllPostList(): Flow<State<List<RealtimeDBModelPostWithRoutine>>>
-}
