+++ app/src/main/java/com/example/shareroutine/di/PostDataSourceModule.kt
@@ -1,9 +1,7 @@
 package com.example.shareroutine.di
 
 import com.example.shareroutine.data.source.PostDataSource
-import com.example.shareroutine.data.source.PostWithRoutineDataSource
 import com.example.shareroutine.data.source.realtime.PostDataSourceImplWithRealtime
-import com.example.shareroutine.data.source.realtime.PostWithRoutineDataSourceImplWithRealtime
 import com.google.firebase.database.DatabaseReference
 import dagger.Module
 import dagger.Provides
@@ -17,17 +15,11 @@ object PostDataSourceModule {
 
     @Singleton
     @Provides
-    fun providePostDataSourceImplWithRealtime(@PostDatabaseRef dbRef: DatabaseReference): PostDataSource {
-        return PostDataSourceImplWithRealtime(dbRef)
-    }
-
-    @Singleton
-    @Provides
     fun providePostWithDataSourceImplWithRealtime(
         @TopDatabaseRef topDbRef: DatabaseReference,
         @PostDatabaseRef postDbRef: DatabaseReference,
         @RoutineDatabaseRef routineDbRef: DatabaseReference
-    ): PostWithRoutineDataSource {
-        return PostWithRoutineDataSourceImplWithRealtime(topDbRef, postDbRef, routineDbRef)
+    ): PostDataSource {
+        return PostDataSourceImplWithRealtime(topDbRef, postDbRef, routineDbRef)
     }
 }
